Import Domain.User instead of Models.User
Change int to long for user id
Use @AuthenticationPrincipal to get current user id in Cart view
Delete Models.User
Comment out UserSeeder
Use lombok to simplify Event and Transaction
Add "addToCart" to SecurityConfiguration to avoid NullPointerException
Add fixes to gallery and fragments views

+++ src/main/java/us/team7pro/EventTicketsApp/Repositories/TransactionRepository.java
@@ -6,8 +6,8 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import us.team7pro.EventTicketsApp.Models.Transaction;
 
 public interface TransactionRepository extends JpaRepository<Transaction, Integer> {
-    public List<Transaction> findByUserID(int userID);
-    public List<Transaction> findByUserIDAndStatus(int userID, boolean status);
-    public Transaction findByUserIDAndEventID(int userID, int eventID);
+    public List<Transaction> findByUserID(long userID);
+    public List<Transaction> findByUserIDAndStatus(long userID, boolean status);
+    public Transaction findByUserIDAndEventID(long userID, int eventID);
     public Transaction findByTransactionID(int transactionID);
 }
