+++ src/main/java/us/team7pro/EventTicketsApp/Services/TransactionService.java
@@ -5,10 +5,10 @@ import us.team7pro.EventTicketsApp.Models.Transaction;
 import java.util.List;
 
 public interface TransactionService {
-    List<Transaction> findByUserID(int userID);
-    List<Transaction> findByUserIDAndStatus(int userID, boolean status);
-    Transaction findByUserIDAndEventID(int userID, int eventID);
+    List<Transaction> findByUserID(long userID);
+    List<Transaction> findByUserIDAndStatus(long userID, boolean status);
+    Transaction findByUserIDAndEventID(long userID, int eventID);
     Transaction findByTransactionID(int transactionID);
     void delete(Transaction t);
-    void addToCart(int userID, int eventID);
+    void addToCart(long userID, int eventID);
 }
